requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
python-dotenv
fastapi
uvicorn
